BREAKING(runtime): remove the boot script
