+++ runtime/package.json
@@ -21,10 +21,6 @@
       "import": "./client/index.js",
       "types": "./client/index.d.ts"
     },
-    "./boot": {
-      "import": "./client/boot.js",
-      "types": "./client/boot.d.ts"
-    },
     "./utils": {
       "import": "./client/utils.js",
       "types": "./client/utils.d.ts"
