+++ runtime/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/boot.ts", "src/utils.ts"],
+  entry: ["src/index.ts", "src/utils.ts"],
   outDir: "client",
   clean: true,
   format: "esm",
